requirement.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
werkzeug==2.3.7
face_recognition==1.3.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
python-dotenv==1.0.0
dlib==19.24.2
cmake==3.27.7
